==> notebooks/02_Model_Selection.ipynb <==
# CUNEF MUCD 2021/2022
## News Classification
Autor:  
- Antonio Tello Gómez

# 2. Model Selection

#Basics
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle


#Functionalities
from collections import Counter
import sys, os
import warnings
warnings.filterwarnings('ignore')

#NLP
import string
import re
import nltk


# Machine Learning
from sklearn.feature_extraction.text import TfidfVectorizer 
from sklearn.pipeline import Pipeline 
from sklearn.model_selection import train_test_split

#Metrics
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score
from sklearn.metrics import roc_auc_score

# Custom Transformer
sys.path.append(os.path.abspath('..'))
sys.path.append(os.path.abspath('../src'))
from src.Preprocessor import TextPreprocessor

# Models
from sklearn.linear_model import LogisticRegression
from sklearn.dummy import DummyClassifier
from lightgbm import LGBMClassifier
from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import Perceptron
from sklearn.linear_model import PassiveAggressiveClassifier
from sklearn.linear_model import SGDClassifier

## Load the Data

#Load the clean data 
#(However, we will use the original data to test the custom transformer)    
df = pd.read_csv('../data/fake_or_real_news_clean.csv')

## Train Test Split

# Train-test split
xtrain, xtest, ytrain, ytest = train_test_split(df['full_text'], df['label'], test_size=0.25, random_state=2022, stratify=df['label'])

# Save train in train folder
xtrain.to_csv('../data/train/X_train.csv', index=False)
ytrain.to_csv('../data/train/y_train.csv', index=False, header=['label'])

# Save test in test folder 
xtest.to_csv('../data/test/X_test.csv', index=False)
ytest.to_csv('../data/test/y_test.csv', index=False, header=['label'])

## Base Model

The base model is a dummy model that always predicts the same value. i.e. all news are real.  
It is used to compare the performance of the other models.

base_model = Pipeline([
    ("preprocessor", TextPreprocessor()),
    ("vectorizer", TfidfVectorizer()),
    ("clf", DummyClassifier(strategy="most_frequent"))
])

base_model.fit(xtrain, ytrain)

base_model.score(xtest, ytest)

ypred = base_model.predict(xtest)
ypred_proba = base_model.predict_proba(xtest)

Function to evaluate the model using popular metrics:  
ROC-AUC,Accuracy,Precision,Recall,F1-score,Confusion matrix...

def evaluate_model(ytest, ypred, ypred_proba = None):
    if ypred_proba is not None:
        print('ROC-AUC score of the model: {}'.format(roc_auc_score(ytest, ypred_proba[:, 1])))
    print('Accuracy of the model: {}\n'.format(accuracy_score(ytest, ypred)))
    print('Classification report: \n{}\n'.format(classification_report(ytest, ypred)))
    print('Confusion matrix: \n{}\n'.format(confusion_matrix(ytest, ypred)))

evaluate_model(ytest, ypred, ypred_proba)

The accuracy of the base model is 0.5


## Training and Model Evaluation

Now that we have a baseline we can start to compare the performance of the other models.

We are going to use the following models:

classifiers =[
    LogisticRegression(),
    MultinomialNB(),
    Perceptron(),
    PassiveAggressiveClassifier(),
    SGDClassifier(),
    RandomForestClassifier(),
    XGBClassifier(),
    LGBMClassifier()
    ]

This loop fits the models in the train data and then generates a report with the metrics in the test data.

%%time
np.random.seed(2022) 
for classifier in classifiers:
    
    # Pipeline
    model = Pipeline(steps=[
        ("preprocessor", TextPreprocessor()),
        ("vectorizer", TfidfVectorizer()),
        ("clf", classifier)]) 
        
    # Fit And Predict Model
    model = model.fit(xtrain, ytrain)
    ypred = model.predict(xtest)
    pickle.dump(model, open('../models/model_selection/' + classifier.__class__.__name__ + '.pkl', 'wb'))
    print('Model Saved')
    # Print Metrics
    print(classifier)
    evaluate_model(ytest, ypred)
    print('------------------------------------------------------')

## Model Selection


Choosing a model is not a trivial choice and it will mostly depend on the particular problem. In a classification problem accuracy on the test set is the most popular metric to assess the performance of a model.  
In our case, we have a balanced dataset, we don't have any particular requirements,  of cost-sensitivity, time, or computational resources and we do not have to deploy our model. Therefore I will use accuracy to evaluate the best model.  
Nevertheless, we can imagine a scenario in which we were developing a real tool or app to detect fake news. In that case, we might want to detect as many fake news as possible even if we classify some real news as fake or we might have a hard time creating a balanced dataset due to most of the news are real. In such cases, metrics like accuracy would not be very useful.


### Winners
Winners: **SGDClassifier and LightGBM**  
Special Mention: Logistic Regression, Passive Aggressive Classifier  
In the following notebook we will optimize the hyperparameters of the winners.
